Initial implementation of zero digit handling
Zero complicates matters since it is not, on its own, a valid lotto number. This initial implementation is fairly naive in that it will not try any additional measures if an input string runs out of characters to complete the pick as it skips over zeros. This means that it will throw an error when it could potentially have rolled backward to try a different combination of digits. The handling for such an edge case would add significant complexity to the code, so it has not been included for now. It may come in a subsequent revision if necessary.